move all code under src folder
