--- apps/web/shared/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useCanvasDragEvent } from './useCanvasDragEvent';
-export { useCanvasDblClickEvent } from './useCanvasDblClickEvent';
