+++ apps/web/src/components/canvas/shapes/PostIts.tsx
@@ -1,0 +1,23 @@
+import { IPostIt } from 'web/src/shared/types';
+import { PostIt } from 'web/src/components/canvas/shapes';
+
+interface Props {
+  postIts: IPostIt[];
+}
+
+export function PostIts({ postIts }: Props) {
+  return (
+    <>
+      {postIts.map(postIt => (
+        <PostIt
+          key={postIt.id}
+          {...{
+            postIt,
+            onDragStart: () => {},
+            onDragEnd: () => {},
+          }}
+        />
+      ))}
+    </>
+  );
+}
