+++ apps/web/src/pages/_app.tsx
@@ -2,7 +2,6 @@ import type { AppProps } from 'next/app';
 import { createGlobalStyle } from 'styled-components';
 import reset from 'styled-reset';
 import { RecoilRoot } from 'recoil';
-import RecoilNexus from 'recoil-nexus';
 
 const GlobalStyle = createGlobalStyle`
   ${reset}
@@ -14,7 +13,6 @@ const GlobalStyle = createGlobalStyle`
 function App({ Component, pageProps }: AppProps) {
   return (
     <RecoilRoot>
-      <RecoilNexus />
       <GlobalStyle />
       <Component {...pageProps} />
     </RecoilRoot>
