+++ apps/web/src/pages/rooms/[roomId].tsx
@@ -1,9 +1,9 @@
 import dynamic from 'next/dynamic';
 import { useState } from 'react';
-import { ActionBar } from 'web/components/layout/ActionBar';
-import { EditModal } from 'web/components/layout/EditModal';
+import { ActionBar } from 'web/src/components/layout/ActionBar';
+import { EditModal } from 'web/src/components/layout/EditModal';
 
-const MainCanvas = dynamic(() => import('web/components/canvas/MainCanvas'), {
+const MainCanvas = dynamic(() => import('web/src/components/canvas/MainCanvas'), {
   ssr: false,
   loading: () => <p>LOADING...</p>,
 });
